Fix problem with multiple handles and uiomux_sleep.
The uio device driver in the kernel stores the last number of irqs
received. This is updated on open and read and is used to determine
when blocked reads should return.

If >1 clients open UIOMux handles and operate in parallel, calls to
uiomux_sleep can return before an interrupt has occurred as the
stored number of irqs received can be different to the current number.

This patch fixes this by performing a non-blocking read during
uiomux_lock to update the stored number of irqs.