Use uio device_index instead of resid to access memory allocation map
The uio.device_index is unique for each UIO device, whereas the resid
is defined by the order in which the UIO devices are opened via
uio_open_named or uio_open_blocks and may differ from one process
to another.

With the addition of this patch the resid argument to uio_malloc()
and uio_free() has become unnecessary and so has been removed.