uiomux_mlock/mtrylock/munlock: Add ability to specify a region of iomem to use
This adds three new apis, uiomux_mlock(), uiomux_mtrylock(), and
uiomux_munlock() to specify a region of iomem to use.

uiomux_mlock() shall be similar to uiomux_malloc(), except that this
allows the caller to specify a region of iomem to use. If a specified
region overlaps other locked (used) regions, the calling thread shall
block until the region becomes available.

uiomux_mtrylock() shall be equivalent to uiomux_mlock(), except that
if the specified region is currently locked, the call shall return
immediately.

uiomux_munlock() shall release the region.

These functions could be useful in the following cases.
* To exclude partial regions which shall be used by others, such as
  kernel drivers or (hard-coded) binary blobs, from the iomem region
  managed by uiomux.
* To share a region exclusively between multiple tasks.