Resolve memory leak in uio device enumeration code
The uio device enumeration code, which scans the uio directory using
scandir(), introduced a memory leak. scandir() allocates buffers with
malloc() and the caller is responsible for releasing the buffers,
which was not being done.
This patch removes the scandir() call and simply enumerates the
available uio devices by trying to open each one in sequence.  The
first device to fail will end the enumeration, so there can be no gaps
in the uio device numbering. i.e. uio0 ~ uioN must all be present.