Determine UIO buffer size at runtime
Previously the maximum amount of memory available to be exported
by each uio device was fixed at compilation time.
This value was used to create the memory management buffer for all
devices when uio_open() was called.
This commit changes the algorithm to allocate the memory management
buffer for each device as it is opened.